requirements.txt
banana-dev==6.3.0 
blinker==1.8.2 
cachetools==5.3.3 
certifi==2024.7.4
cffi==1.16.0 
charset-normalizer==3.3.2 
click==8.1.7 
cmake==3.30.0 
filelock==3.15.4 
Flask==3.0.3 
Flask-Cors==4.0.1 
fsspec==2024.6.1 
google-api-core==2.19.1 
google-auth==2.31.0 
google-cloud-core==2.4.1
google-cloud-storage==2.17.0 
google-crc32c==1.5.0 
google-resumable-media==2.7.1 
googleapis-common-protos==1.63.2 
huggingface-hub==0.23.4 
idna==3.7 
itsdangerous==2.2.0 
Jinja2==3.1.4
-e git+https://github.com/cheind/py-lapsolver.git@347b73f7216dd39e3689f2f64827d45141250962#egg=lapsolver 
MarkupSafe==2.1.5 
mpmath==1.3.0 
networkx==3.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1 
nvidia-cuda-cupti-cu12==12.1.105 
nvidia-cuda-nvrtc-cu12==12.1.105 
nvidia-cuda-runtime-cu12==12.1.105 
nvidia-cudnn-cu12==8.9.2.26 
nvidia-cufft-cu12==11.0.2.54 
nvidia-curand-cu12==10.3.2.106 
nvidia-cusolver-cu12==11.4.5.107 
nvidia-cusparse-cu12==12.1.0.106 
nvidia-nccl-cu12==2.20.5 
nvidia-nvjitlink-cu12==12.5.82 
nvidia-nvtx-cu12==12.1.105 
packaging==24.1 
pedalboard==0.9.9 
praat-parselmouth==0.4.3 
proto-plus==1.24.0 
protobuf==5.27.2 
psola==0.0.1 
pyasn1==0.6.0 
pyasn1_modules==0.4.0 
pycparser==2.22 
pypar==0.0.6 
python-dotenv==1.0.1 
PyYAML==6.0.1 
regex==2024.5.15 
requests==2.32.3 
resemble==1.4.3 
rsa==4.9 
safetensors==0.4.3 
scipy==1.14.0 
sentencepiece==0.2.0 
setuptools==70.2.0 
soundfile==0.12.1 
sympy==1.12.1 
tokenizers==0.19.1 
torch==2.3.1 
tqdm==4.66.4 
transformers==4.42.3 
typing_extensions==4.12.2 
urllib3==2.2.2
Werkzeug==3.0.3